back/requirements.txt
Django==2.2
djangorestframework==3.8.2
django-cors-headers==2.4.0
mysqlclient==1.3.13
Pillow==5.2.0
requests==2.23.0
uWSGI==2.0.17.1